// src/driver-portal/dto/accept-invite.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsString, IsNotEmpty, IsUUID } from 'class-validator';

export class AcceptInviteDto {
  @ApiProperty({
    description: 'Invite token generated by fleet manager',
    example: '550e8400-e29b-41d4-a716-446655440000',
  })
  @IsString()
  @IsNotEmpty()
  @IsUUID('4')
  token: string;
}

export class AcceptInviteResponseDto {
  @ApiProperty()
  driver_id: string;
}